requirements.txt
# requirements.txt

nextcord
requests
python-dotenv
colorama
humanfriendly
